# Use new fsspec.parquet module to accelerate reads from remote storage