--- a/tests/unit/ops/test_categorify.py
+++ b/tests/unit/ops/test_categorify.py
@@ -66,12 +66,17 @@ def test_categorify_size(tmpdir, cpu, include_nulls):
             if size
         }
     else:
+        # Ignore first element if it is NaN
+        if vocab["session_id"].iloc[:1].isna().any():
+            session_id = vocab["session_id"].iloc[1:]
+            session_id_size = vocab["session_id_size"].iloc[1:]
+        else:
+            session_id = vocab["session_id"]
+            session_id_size = vocab["session_id_size"]
         expected = dict(zip(vals.index.values_host, vals.values_host))
         computed = {
             session: size
-            for session, size in zip(
-                vocab["session_id"].values_host, vocab["session_id_size"].values_host
-            )
+            for session, size in zip(session_id.values_host, session_id_size.values_host)
             if size
         }
     first_key = list(computed.keys())[0]