This is a follow-up to #1239 to add the necessary logic to leverage the new `fsspec.parquet` module to improve read performance from remote storage (for cudf>=21.12). The full functionality in this PR requires the latest (unreleased) versions of cudf and fsspec to become active. However, the performance motivation is pretty clear:

**Before this PR (Including the #1239 fix)**:

```python
In [1]: import nvtabular as nvt

In [2]: %time ds = nvt.Dataset("gs://workshop-datasets/criteo-parque", engine="parquet", use_fsspec_parqu
   ...: et=False)
CPU times: user 2.57 s, sys: 2.44 s, total: 5.01 s
Wall time: 5.63 s

In [3]: %time ds.head()
CPU times: user 5.05 s, sys: 3.15 s, total: 8.19 s
Wall time: 8.73 s
Out[5]: 
   label    I1   I2    I3    I4  ...          C22         C23         C24         C25         C26
0      1     5  110  <NA>    16  ...    632402057  -675152885  2091868316   809724924  -317696227
1      0    32    3     5  <NA>  ...   -924717482   391309800  1966410890 -1726799382 -1218975401
2      0  <NA>  233     1   146  ...    632402057  -675152885   883538181   -10139646  -317696227
3      0  <NA>   24  <NA>    11  ...         <NA>  -601842477 -1196941392   809724924 -1218975401
4      0    60  223     6    15  ...  -1466264923 -1441566359  -434539461  -842849922  -317696227

[5 rows x 40 columns]
```
Where both the `Dataset` and `head()` operations utilize ~13GB of host memory.

**After this PR**:

```python
In [1]: import nvtabular as nvt

In [2]: %time ds = nvt.Dataset("gs://workshop-datasets/criteo-parque", engine="parquet")
CPU times: user 1.2 s, sys: 574 ms, total: 1.77 s
Wall time: 2.6 s

In [3]: %time ds.head()
CPU times: user 2.92 s, sys: 1.53 s, total: 4.46 s
Wall time: 4.97 s
Out[3]: 
   label    I1   I2    I3    I4  ...          C22         C23         C24         C25         C26
0      1     5  110  <NA>    16  ...    632402057  -675152885  2091868316   809724924  -317696227
1      0    32    3     5  <NA>  ...   -924717482   391309800  1966410890 -1726799382 -1218975401
2      0  <NA>  233     1   146  ...    632402057  -675152885   883538181   -10139646  -317696227
3      0  <NA>   24  <NA>    11  ...         <NA>  -601842477 -1196941392   809724924 -1218975401
4      0    60  223     6    15  ...  -1466264923 -1441566359  -434539461  -842849922  -317696227

[5 rows x 40 columns]
```
Where the `Dataset` and `head()` operations only utilize ~3GB of host memory.

**Important Note**: The "before" numbers assume that you are using the fix in #1239. The current `main` behavior is actually ~50% slower and utilizes ~26GB of host memory.

<!--

Thank you for contributing to NVTabular :)

Here are some guidelines to help the review process go smoothly.

1. Please write a description in this text box of the changes that are being
   made.

2. Please ensure that you have written units tests for the changes made/features
   added.

3. If you are closing an issue please use one of the automatic closing words as
   noted here: https://help.github.com/articles/closing-issues-using-keywords/

4. If your pull request is not ready for review but you want to make use of the
   continuous integration testing facilities please label it with `[WIP]`.

5. If your pull request is ready to be reviewed without requiring additional
   work on top of it, then remove the `[WIP]` label (if present) and replace
   it with `[REVIEW]`. If assistance is required to complete the functionality,
   for example when the C/C++ code of a feature is complete but Python bindings
   are still required, then add the label `[HELP-REQ]` so that others can triage
   and assist. The additional changes then can be implemented on top of the
   same PR. If the assistance is done by members of the rapidsAI team, then no
   additional actions are required by the creator of the original PR for this,
   otherwise the original author of the PR needs to give permission to the
   person(s) assisting to commit to their personal fork of the project. If that
   doesn't happen then a new PR based on the code of the original PR can be
   opened by the person assisting, which then will be the PR that will be
   merged.

6. Once all work has been done and review has taken place please do not add
   features or make changes out of the scope of those requested by the reviewer
   (doing this just add delays as already reviewed code ends up having to be
   re-reviewed/it is hard to tell what is new etc!). Further, please do not
   rebase your branch on master/force push/rewrite history, doing any of these
   causes the context of any comments made by reviewers to be lost. If
   conflicts occur against master they should be resolved by merging master
   into the branch used for making the pull request.

Many thanks in advance for your cooperation!

-->